refactor(tests): split kit and bench out of integration/ into tests/kit and tests/bench
