+++ tests/integration/edge/edge_test.go
@@ -12,7 +12,7 @@ import (
 
 	dto "github.com/rabbytesoftware/quiver/internal/api/v0/dto"
 	"github.com/rabbytesoftware/quiver/internal/domain"
-	"github.com/rabbytesoftware/quiver/tests/integration/kit"
+	"github.com/rabbytesoftware/quiver/tests/kit"
 )
 
 func TestMain(m *testing.M) { kit.Main(m) }
