+++ tests/kit/bench.go
