+++ tests/kit/client.go
