+++ tests/kit/oracle.go
