+++ tests/kit/suite.go
